# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Grow5X API
  description: |
    API del backend de Grow5X - Plataforma de trading e inversiones con sistema de referidos y comisiones.

    ## Autenticación
    La mayoría de endpoints requieren autenticación mediante JWT token en el header Authorization:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Roles
    - **user**: Usuario regular con acceso a funciones básicas
    - **admin**: Administrador con acceso a funciones administrativas

    ## Códigos de Error Comunes
    - `VALIDATION_ERROR`: Error de validación de datos
    - `UNAUTHORIZED`: Token inválido o faltante
    - `FORBIDDEN`: Permisos insuficientes
    - `RATE_LIMIT_EXCEEDED`: Límite de requests excedido
    - `INTERNAL_ERROR`: Error interno del servidor
  version: 1.0.0
  contact:
    name: Grow5X Team
    email: support@grow5x.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.grow5x.app
    description: Servidor de producción
  - url: http://localhost:5000
    description: Servidor de desarrollo
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Inicio de sesión
      description: Autentica un usuario y devuelve JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@ejemplo.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Cerrar sesión
      description: Invalida el token JWT actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Sesión cerrada exitosamente
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Obtener información del usuario actual (alias legacy)
      description: Devuelve la información del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Registro de nuevo usuario
      description: Registra un nuevo usuario con código de referido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - referralCode
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@ejemplo.com
                password:
                  type: string
                  minLength: 8
                  example: password123
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: Juan
                  description: Opcional - Nombre del usuario
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: Pérez
                  description: Opcional - Apellido del usuario
                phone:
                  type: string
                  minLength: 10
                  maxLength: 20
                  example: '+1234567890'
                  description: Opcional - Teléfono del usuario
                country:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: México
                  description: Opcional - País del usuario
                referralCode:
                  type: string
                  minLength: 1
                  example: REF123456
                telegramUsername:
                  type: string
                  example: '@usuario'
                  description: Opcional - Username de Telegram para notificaciones
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error de validación o email ya existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/me:
    get:
      tags:
        - User Profile
      summary: Obtener información del usuario actual (endpoint canónico)
      description: Devuelve la información del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido del endpoint de login.
        Formato: `Bearer <token>`
  schemas:
    User:
      type: object
      properties:
        userId:
          type: string
          example: USR_20240101_001
        email:
          type: string
          format: email
          example: usuario@ejemplo.com
        firstName:
          type: string
          example: Juan
        lastName:
          type: string
          example: Pérez
        phone:
          type: string
          example: '+1234567890'
        country:
          type: string
          example: México
        role:
          type: string
          enum:
            - user
            - admin
          example: user
        referralCode:
          type: string
          example: REF123456
        isActive:
          type: boolean
          example: true
        isVerified:
          type: boolean
          example: true
        balances:
          type: object
          properties:
            available:
              type: number
              example: 1500.5
            pending:
              type: number
              example: 200
            total:
              type: number
              example: 1700.5
        referralStats:
          type: object
          properties:
            directReferrals:
              type: integer
              example: 5
            totalReferrals:
              type: integer
              example: 15
            totalCommissions:
              type: number
              example: 750
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login exitoso
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            user:
              $ref: '#/components/schemas/User'
            expiresIn:
              type: string
              example: 24h
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error de validación
        code:
          type: string
          example: VALIDATION_ERROR
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Email inválido
...