# backend/src/models/index.js
/**
 * Models Index
 * Centralized export of all Mongoose models
 */

const User = require('./User');
const Package = require('./Package');
const Wallet = require('./Wallet');
const Purchase = require('./Purchase');
const Transaction = require('./Transaction');
const Commission = require('./Commission');
const BenefitLedger = require('./BenefitLedger');
const BenefitSchedule = require('./BenefitSchedule');
const Ledger = require('./Ledger');
const SpecialCode = require('./SpecialCode');
const Withdrawal = require('./Withdrawal');
const AuditLog = require('./AuditLog');
const JobState = require('./JobState');
const Cohort = require('./Cohort');
const Settings = require('./Settings');

module.exports = {
  User,
  Package,
  Wallet,
  Purchase,
  Transaction,
  Commission,
  BenefitLedger,
  BenefitSchedule,
  Ledger,
  SpecialCode,
  Withdrawal,
  AuditLog,
  JobState,
  Cohort,
  Settings
};